requirements.txt
Django>=4.2
djangorestframework
gunicorn
psycopg2-binary
django-cors-headers
whitenoise
python-decouple
dj-database-url
django-filter


